// plugins/org.eclipse.osee.web/src/app/ple/messaging/shared/constants/AttributeTypeId.enum.ts
/*********************************************************************
 * Copyright (c) 2021 Boeing
 *
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Boeing - initial API and implementation
 **********************************************************************/
export const enum ATTRIBUTETYPEID {
    NAME = "1152921504606847088",
    DESCRIPTION = "1152921504606847090",
    INTERFACENODEBGCOLOR = "5221290120300474048",
    INTERFACENODEADDRESS = "5726596359647826656",
    INTERFACETRANSPORTTYPE="4522496963078776538"
}